refactor: 페이지 반환과 REST 응답 반환 구분
