+++ src/main/java/shopping/controller/CartPageController.java
@@ -1,0 +1,13 @@
+package shopping.controller;
+
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+
+@Controller
+public class CartPageController {
+
+    @GetMapping("/cart")
+    public String renderCartPage() {
+        return "cart";
+    }
+}
