+++ src/main/java/shopping/controller/LoginPageController.java
@@ -4,10 +4,10 @@ import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
 
 @Controller
-public class CartController {
-
-    @GetMapping("/cart")
-    public String renderCartPage() {
-        return "cart";
+public class LoginPageController {
+    @GetMapping("/login")
+    public String renderLoginPage() {
+        return "/login";
     }
+
 }
