+++ src/main/java/shopping/controller/ProductPageController.java
@@ -7,11 +7,11 @@ import shopping.dto.response.ProductResponses;
 import shopping.service.ProductService;
 
 @Controller
-public class ProductController {
+public class ProductPageController {
 
     private final ProductService productService;
 
-    public ProductController(final ProductService productService) {
+    public ProductPageController(final ProductService productService) {
         this.productService = productService;
     }
